server_config/requirements.txt
PyYAML>=6.0
python-iptables>=1.0.0
dnspython>=2.2.1